+++ mlxtend/classifier/tests/test_logistic_regression.py
@@ -21,7 +21,10 @@ X[:, 1] = (X[:, 1] - X[:, 1].mean()) / X[:, 1].std()
 
 def test_logistic_regression_gd():
     t = np.array([0.51, 1.18, 4.40])
-    lr = LogisticRegression(epochs=100, eta=0.01, learning='gd', random_seed=0)
+    lr = LogisticRegression(epochs=100,
+                            eta=0.01,
+                            minibatches=1,
+                            random_seed=0)
 
     lr.fit(X, y)  # 0, 1 class
     np.testing.assert_almost_equal(lr.w_, t, 2)
@@ -30,8 +33,10 @@ def test_logistic_regression_gd():
 
 def test_logistic_regression_sgd():
     t = np.array([0.50, 1.16, 4.38])
-    lr = LogisticRegression(epochs=100, eta=0.01,
-                            learning='sgd', random_seed=0)
+    lr = LogisticRegression(epochs=100,
+                            eta=0.01,
+                            minibatches=len(y),
+                            random_seed=0)
 
     lr.fit(X, y)  # 0, 1 class
     np.testing.assert_almost_equal(lr.w_, t, 2)
@@ -41,7 +46,7 @@ def test_logistic_regression_sgd():
 def test_l2_regularization_gd():
     lr = LogisticRegression(eta=0.01,
                             epochs=20,
-                            learning='gd',
+                            minibatches=1,
                             l2_lambda=1.0,
                             regularization='l2',
                             random_seed=0)
@@ -55,15 +60,17 @@ def test_l2_regularization_gd():
 
 
 def test_l2_regularization_sgd():
-    lr = LogisticRegression(eta=0.01, epochs=20,
-                            learning='sgd',
+    lr = LogisticRegression(eta=0.01,
+                            epochs=100,
+                            minibatches=len(y),
                             l2_lambda=1.0,
                             regularization='l2',
                             random_seed=0)
     lr.fit(X, y)
     y_pred = lr.predict(X)
-    expect_weights = np.array([0.09,  0.232,  0.35])
+    expect_weights = np.array([0.,  0.24,  0.35])
 
     np.testing.assert_almost_equal(lr.w_, expect_weights, 2)
     acc = sum(y_pred == y) / float(len(y))
-    assert(acc == 1.0)
+
+    assert(acc == 0.97)