+++ mlxtend/regressor/tests/test_linear_regression.py
@@ -26,19 +26,19 @@ expect_rm_lstat_std = np.array([0.000, 0.389, -0.499])
 
 
 def test_univariate_normal_equation():
-    ne_lr = LinearRegression(solver='normal equation')
+    ne_lr = LinearRegression(minibatches=None)
     ne_lr.fit(X_rm, y)
     assert_almost_equal(ne_lr.w_, expect_rm, decimal=3)
 
 
 def test_univariate_normal_equation_std():
-    ne_lr = LinearRegression(solver='normal equation')
+    ne_lr = LinearRegression(minibatches=None)
     ne_lr.fit(X_rm_std, y_std)
     assert_almost_equal(ne_lr.w_, expect_rm_std, decimal=3)
 
 
 def test_univariate_gradient_descent():
-    gd_lr = LinearRegression(solver='gd',
+    gd_lr = LinearRegression(minibatches=1,
                              eta=0.001,
                              epochs=500,
                              random_seed=0)
@@ -47,7 +47,7 @@ def test_univariate_gradient_descent():
 
 
 def test_univariate_stochastic_gradient_descent():
-    sgd_lr = LinearRegression(solver='sgd',
+    sgd_lr = LinearRegression(minibatches=len(y),
                               eta=0.0001,
                               epochs=100,
                               random_seed=0)
@@ -56,13 +56,16 @@ def test_univariate_stochastic_gradient_descent():
 
 
 def test_multivariate_normal_equation():
-    ne_lr = LinearRegression(solver='normal equation')
+    ne_lr = LinearRegression(minibatches=None)
     ne_lr.fit(X_rm_lstat, y)
     assert_almost_equal(ne_lr.w_, expect_rm_lstat, decimal=3)
 
 
 def test_multivariate_gradient_descent():
-    gd_lr = LinearRegression(eta=0.001, epochs=500, solver='gd', random_seed=0)
+    gd_lr = LinearRegression(eta=0.001,
+                             epochs=500,
+                             minibatches=1,
+                             random_seed=0)
     gd_lr.fit(X_rm_lstat_std, y_std)
     assert_almost_equal(gd_lr.w_, expect_rm_lstat_std, decimal=3)
 
@@ -70,7 +73,7 @@ def test_multivariate_gradient_descent():
 def test_multivariate_stochastic_gradient_descent():
     sgd_lr = LinearRegression(eta=0.0001,
                               epochs=500,
-                              solver='sgd',
+                              minibatches=len(y),
                               random_seed=0)
     sgd_lr.fit(X_rm_lstat_std, y_std)
     assert_almost_equal(sgd_lr.w_, expect_rm_lstat_std, decimal=2)